refactor: nicknameValidator member utility로 분리
