+++ src/components/UserRegister/Fields/Nickname.tsx
@@ -14,24 +14,13 @@ import { Alert } from '~/components/ModalContent';
 import Question from '~/components/UserRegister/Question';
 import {
   createRandomNickname,
+  nicknameValidator,
   useUpdateMyInfoFormContext,
   useValidateNickname,
 } from '~/services/member';
 import { flex, palettes } from '~/styles/utils';
 
 const fieldName = 'nickname';
-const isValidLength = (value: string) => {
-  const { length } = value;
-  return 1 <= length && length <= 11;
-};
-
-const isValidWhiteSpace = (value: string) => {
-  return !/(^\s|\s$|\s{2,})/.test(value);
-};
-
-const isValidFormat = (value: string) => {
-  return !/[^\w 가-힣]/.test(value);
-};
 
 const Nickname = () => {
   const [submitModalOpen, setSubmitModalOpen] = useState(false);
@@ -118,20 +107,7 @@ const Nickname = () => {
           autoComplete="off"
           id={nicknameFieldId}
           {...register(fieldName, {
-            validate: (value) => {
-              if (!isValidLength(value)) {
-                return '길이는 1 ~ 11 사이여야 합니다.';
-              }
-
-              if (!isValidWhiteSpace(value)) {
-                return '공백문자를 연달아 사용하거나, 처음과 마지막에 사용할 수 없습니다.';
-              }
-
-              return (
-                isValidFormat(value) ||
-                '한글, 영문, 숫자, _, 공백만 사용할 수 있습니다.'
-              );
-            },
+            validate: nicknameValidator,
           })}
         />
         {errorMessage && <AlertText>{errorMessage}</AlertText>}
