+++ src/services/member/utils/index.ts
@@ -1,3 +1,4 @@
-export * from './role';
 export * from './createRandomNickname';
+export * from './nicknameValidator';
+export * from './role';
 export * from './types';
